Stylistic refactor Fixes lint issues introduced by https://github.com/nickbabcock/jomini/pull/27
